Add files via upload
Consensus means the interaction between groups of agents in a team, via sensors or a communication network, to reach an agreement on a common value or state (Olfati-Saber and Shamma 2005; Schenato and Gamba 2007; Olfati-Saber et al. 2006).Its purpose is to synchronize the states or outputs of the systems under the designed distributed control protocols.